MB-31664: Don't accept streamRequests when receiving first disk snapshot
With the plan to remove the dedicated backfill queue, we want remove the
reliance on checkpointID of zero meaning that we are receiving the
initial disk snapshot.  Also vbuckets currently have a backfillphase
flag which is used to indicate whether an item should be added to the
backfill queue or not.

To keep functionality unchanged after removing the dedicated backfill
queue, we want to temporarily fail a stream request from a vbucket
whilst it is receiving its initial disk snapshot.  Therefore we
introduce a receivingInitialDiskSnapshot flag which is used solely for
this reason.

Change-Id: Ia0b67bac70d84d9879ddb3d87d72f78e6ea3326a
Reviewed-on: http://review.couchbase.org/100913
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>