Request RBAC data for external auth if cached entry is old
Request a fresh copy of the RBAC data for the user if the cached
entry is older than 2x the push interval for user in the RBAC
cache.

The motivation for this is if an external user disconnects
and stays offline for a day or two we don't want the user
to be able to use the old privilege set when he logs back
in until ns_server gets around pushing the updated set.

Change-Id: I86945dc3df87c1e22cce0acd533c176df4aff75b
Reviewed-on: http://review.couchbase.org/100918
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>