Inline conn_ship_log in statemachine.cc
Now that we've moved all of the message producers as part of
the connection object the remainder of the method is relatively
small so it may be moved into the method in statemachine.cc
(which makes it easier for further cleanup).

In addition the comments was updated to no longer refer
to the old (and removed) return value ENGINE_WANTS_MORE.

And we're not keeping the cookie in an ewouldblock state
"forever".

Change-Id: I7ffa4c7432e22aac5636a0e47ab375c5e3fbccb6
Reviewed-on: http://review.couchbase.org/100951
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>