MB-31731: Only push config value to a running ext auth mgd
The configuration file is read _before_ the external auth
manager is started so the config value cannot be pushed
to the object.

The original unit test cases did _not_ provide a value
for this parameter (and used the default instead), but
utilized the reconfigure method to set the interval
to a low value to use during the test. This patch
sets the value to 30 minutes to make sure that we
_do_ handle the case where it present in the configuration
when no auth manager is running.

Change-Id: I08cc516ec67bfeadb1fafbcb60596331d3f41c52
Reviewed-on: http://review.couchbase.org/100992
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>