MB-31718: Correct collection doc counts after rollback
We currently do not track collection diskCounts in the event
of a rollback. In particular, as we do not increment the
collection diskCount when we rollback a deletion, we can crash
if we perform deletions until the diskCount would go below 0.

Update EPDiskRollbackCB to both increment and decrement the
collection document diskCount when we rollback sets and deletions
respectively.

Change-Id: I02fa627d7cf2044a70700dc5dd6a5a78be355417
Reviewed-on: http://review.couchbase.org/100972
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>