MB-31739: [2/2] Update create/drop system events per design changes
The create event needs to transmit
* manifest uid
* collection uid
* scope uid
* collection ttl (which is optional)
* collection name

The drop event needs only to transmit
* manifest uid
* collection uid

A DCP consumer can then correctly track the entire collection
manifest using this data.

The optional max_ttl is handled by having different versions of the
create collection event, so we can cleanly trasmit the different
sized structures.

These updates are achieved by updating the flatbuffer schema and
allowing the extra data (name and ttl) to be passed through
existing layers that join collections changes to DCP.

Note that we are storing and transmitting a maxTtl, but the Manifest
does not yet accept a max_ttl field and the max_ttl values themselves
aren't applied to Items.

Change-Id: I0f66aac3000b5d95ac8cf2997e00feb6a6fce415
Reviewed-on: http://review.couchbase.org/101015
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>