Refactor: Use a refcount on a per cookie base
Previously the underlying engine could "reserve" a cookie, which
should be used by an engine if it held a reference to the cookie
in the engine and would operate on behalf of that cookie at a
later time. That would guarantee that the core wouldn't invalidate
the cookie.

This worked very well in the old design where each connection
would only have a single cookie, and the cookie had the same
lifetime as the connection.

In OOO mode each connection may have multiple cookies at the
same time, and we need to be able to distinguish which cookie
we're trying to reserve.

It is beyond the scope of this patch to refactor the interface
between the core and the engine to introduce shared_ptr's or
other mechanisms. That would grow the scope of the entire
task, and we don't know if it'll introduce a performance
overhead etc. It may be refactored in a dedicated patch at
a later time.

Change-Id: Ifa6d63e3dc05887f504b5d3304290cb70728968f
Reviewed-on: http://review.couchbase.org/101007
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>