Fix calculation of sleep time for auth daemon
The External Authentication Thead is supposed to send the active
users to the providers at a fixed (configurable) interval, but
it should also service the normal authentication request.

To avoid running a "busy loop" checking if it is supposed to send
the next notification it sleeps on a condition variable for a
period of time. This was _supposed_ to be when the next push
should be sent, but the calculation was off. Instead of the
correct calculation:

    interval - (now - last sent)

It used

   now - last sent

However the check for if the list should be pushed or not was
correct so we ended up with a something which starts off as
a busy-loop and do backoff. It has the side effect that
the sleep is now longer and longer, but it is still wrong
so we'll be sleeping when we're supposed to send the actual
notification.

If we receive a notification on the condition variable right
before we're supposed to push the list (this could be due to
a normal auth request) we would not push the list, but go
back to sleep for the full duration of a push interval
so the other end would think that we lost one.

Change-Id: I794a07067dfbfa292e8a74741858ee36700743e1
Reviewed-on: http://review.couchbase.org/101051
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>