MB-24546: Make collections-details consistent with vbucket-details
Make the cbstats "collections-details" stats print out NOT_MY_VBUCKET
errors so that it is consistent with the "vbucket-details" command
when run with a vbucket parameter that does not exist on a given node.
Currently when running the "collections-details" command with an
invalid vBucket nothing is printed.

Also, log the exception if one is thrown.

Change-Id: I8f64badcae172f4c006e09aa364b0a681b1ad1fc
Reviewed-on: http://review.couchbase.org/101078
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>