MB-26887: Add new methods to copy or chain data to send pipe [3/7]
Add two new methods to McbpConnection to either copy the data to
the send buffer, or add a reference to the buffer with a cleanup
function to call when libevent is done with the data.

A new SendBuffer interface is added to allow for various buffers
to be chained to the output stream.

Change-Id: I1f396bb905687d4ce7a7e7e5c84b1da3576bc64e
Reviewed-on: http://review.couchbase.org/100030
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>