MB-30487: [2/2] Enforce the collection max_ttl
This patch refactors the existing bucket max_ttl enforcement and embeds
it with the collection VB manifest.

Moving the bucket ttl enforcement to the manifest allows collection and
bucket TTL to be checked with one collections map lookup.

The patch updates the following operations to apply the defined limit

* add
* set
* replace
* GAT
* set-with-meta

In terms of actual enforcement, the design requires that any collection
max_ttl overrides any bucket max_ttl, this logic is in processExpiryTime
and tested in CollectionsExpiryLimitTest::operation_test

Change-Id: I15fb9f3d212b9f7027df1549d2023ff994b46dfa
Reviewed-on: http://review.couchbase.org/101230
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>