MB-30487: [1/2] Accept max_ttl field in collection entry
This commit updates the JSON parsing so that we can
accept a max_ttl field in the collection manifest.

Once a max_ttl is specified it is propagated down
to the VB::Manifest during collection creation.

There is no support for changing an existing max_ttl,
only create a new collection with a max_ttl is
currently supported by this commit.

Change-Id: I13e099bccf794d44000586524689abea4da4af0f
Reviewed-on: http://review.couchbase.org/101229
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>