MB-31797: Correct tracking of vBucket states in vbStateCount
Correct the bugs currently in the code that are found by setting
the CB_DEVELOPMENT_ASSERTS flag to ON.

Currently vBucket state changes are only decremented in the VBucketMap
vbStateCount when the vBucket already exists. As the vbStateCount is
a set of NonNegativeCounters, we must also increment the new state
counter to prevent firing development assertions on following decrements
of the old state.

Change-Id: Ida73bc3e22f88b40caf086569345355fe8fd6250
Reviewed-on: http://review.couchbase.org/101547
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>