MB-31810: Implement collection persisted high seqno
GSI need the ability to get the highest seqno of a collection,
just like they already do for a vBucket. To supply the high seqno
for each collection after restart/warmup without scanning the
entire database we need to persist a high seqno for each collection.

Add the requried code to persist and load the high seqno for each
collection in a similar way to the collection item counts.

In the case of rollbacks, reset the collection stats counter to the
value in the rolled back document.

Change-Id: I9bd397e07e515931cffc37f0f721476610e37037
Reviewed-on: http://review.couchbase.org/101332
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>