MB-31817: Fix collecton completeDeletion inconsistency.
The MB describes how by the use of a system-event (delete hard)
to sync the _local collection metadata to the data of the VB
can lead to an inconsistency under failure, this commit removes
the use of delete hard and uses the compactor to update the
metadata of the new datafile before making it the current datafile.

A more complete description of the updated delete flow is documented
in collections.md and is replicated here:

Deleting a collection begins by the VB::Manifest deleting the
`SystemEvent::Collection` that was originally written when the
collection was created. The seqno of this deletion is recorded in the
collection's entry as the end seqno. The items of the collection still
exist, but now they are termed 'logically deleted'.

The items of the collection are now removed from the system by the
compactor.

The compactor initialises a VB::Manifest from the datafile's _local
manifest, thus the compactor makes decisions about item deletion
consistently with the datafile.

As the kvstore compactor iterates the bySeqno index it will decide if
the key@seqno should be discarded, this is asking it's VB::Manifest if
the item `isLogicallyDeleted`. If a key has a seqno which is in the
deleted range (less than the start or end seqno) it is discarded from
the database.

A collection is considered fully deleted when the compactor finally
encounters the collection's deleted marker, the assumption being we must
of visited all items that could of been in the collection, so the
collection is now empty.

On encountering the deleted marker the compactor updates: * The in-
memory VB::Manifest to erase the collection entry (freeing a small
amount of resource) * The compaction owned VB::Manifest to erase the
collection entry, ready for the final stage

The final stage of compaction is now to update the _local collection
metadata by serialising the compaction owned VB::Manifest into a
flatbuffer format, the serialised data will not contain any of the fully
dropped collections, making the metadata consistent with the data.

The actual delete marker will be removed by tombstone purging (e.g. 3
days later) this ensures any DCP clients which were disconnected can
reconnect to the VB and for 3 days still discover that the collection
was dropped After 3 days, the marker is purged and the purge-seqno
blocks those clients who are now inconsistent with the database.

Change-Id: I854bf61046c36d5e1340c994cf35cfaf80a18718
Reviewed-on: http://review.couchbase.org/101273
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>