MB-31907: [2/2] Persist and replicate scope events
Update the VB::Manifest so that it can queue a system
event when a scope is created or dropped, this system
event is very similar to collection create/drop.

The patch also updates the filter code so that a stream
which is scope filtered closes when the scope is removed.

Change-Id: I42a80e93f0f9caa8707cd9faa3384d44c03cfd2d
Reviewed-on: http://review.couchbase.org/101454
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>