MB-31810: HashTableVisitors HashBucketLock wrapping functions
To implement an in memory collection high seqno we need to set a
variable that will be stored in a Collections ManifestEntry on
every new high senqo for that particular collection. To do so, we
must wrap every operation that may update the high seqno in a
Collections::VB::Manifest::ReadHandle/CachingReadHandle. We should do
so before acquiring any HashBucketLocks to minimise HashTable
contention and abort operations early on non-valid collections.

This presents an interesting case for Ephemeral buckets as the
pageOut() function can modify a collection high seqno (we update the
high seqno on deletion of an item). To prevent lock order inversions
we still need to acquire a Collections::VB::Manifest::ReadHandle
before we acquire the HashBucketLock. As the HashTableVisitors iterate
over HashBuckets and only acquire a HashBucketLock once (at the start
of iteration) we need to modify HashTable::pauseResumeVisit to allow
us to take aquire a Collections::VB::Manifest::ReadHandle lock before
doing so, and dispose of it after the HashBucketLock.

To implement this functionality with requiring that every
HashTableVisitor locks HashBucket visits in a
Collections::VB::Manifest::ReadHandle we can introduce two virtual
functions that a HashTableVisitor can override. These functions,
setUp and tearDown, will allow a visitor to set any required state
or acquire any locks before the HashBucketLock is acquired, and then
dispose of them appropriately after the HashBucketLock has been
released.

To allow such Visitors to avoid memory allocations in acquiring a
Collections::VB::Manifest::ReadHandle for each HashBucket, make the
ReadHandle default constructable in an unlocked state so that we
can assign locked/unlocked ReadHandles as requried.

Currently the PagingVisitor does not use the added readHandle, it was
added to demonstrate functionality, but the use will be added in a
following patch.

Change-Id: I7f06ff5f3f097d6ccc76ef9f020420a866f24802
Reviewed-on: http://review.couchbase.org/102442
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>