MB-31867: ns_server encodes the vbucket state in the value
The set_vbucket command contains the new state in the body
of the message. Instead of encoding the field in the extras
(like "all" other commands) ns_server sets extras to 0 and
put the new state in the value field.

Previously we did not check where the extras was specified,
but as part of adding support for flex extras the code was
rewritten to pick out the extras field and read the data
from there.

Change-Id: I20f9a4719af8d871a6f39693afd11bf7ae8219a8
Reviewed-on: http://review.couchbase.org/101414
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>