getPausedReason to return PausedReason
getPausedReason currently returns a string.  However it is only used
within the ConnHandler::addStats function, therefore the use case can be
easily changed to use the to_string function.

This allows getPausedReason to return the enum type PausedReason which
will be used in a future patch where we want to identify connecions
with a specfic PausedReason.

Change-Id: I5d0265b05d3796f30779857ab70bda38a0f508fa
Reviewed-on: http://review.couchbase.org/101436
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>