MB-30302: Tweak log message verbosity to prevent log spamming
Drop the "Stream closing..." message to info level when we end
with reason OK to prevent views from spamming our log. Also, remove
the duplicated log message.

Move the "CheckpointManager::registerCursorBySeqno" message to the
ActiveStream locations from which it is called so that we can
log it with the connection logger and use the connection loggers
verbosity setting (again to prevent spamming of our logs). Add a
log message to the CheckpointManager constructor so that we still
log the creation of the persistence cursor.

Change-Id: I81159859f0f5b978534d26f8d3316d7da96667d8
Reviewed-on: http://review.couchbase.org/101475
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>