MB-31797: Correct vbStateCount array indexing with offset
Correct the bugs currently in the code that are found by setting
the CB_DEVELOPMENT_ASSERTS flag to ON.

vBStateCount (part of vBucket map) incorrectly maps the
vbucket_state_t enum to an array due to an off by one error. The
vbucket_state_t enum starts at value 1, not 0, so we need to subtract
1 (vbucket_state_active) whenever we use the '[]' operator.

Currently we overrun the bounds of memory allocated to the array by
2 bytes. As the array is the last member of the class and the class is
probably some odd number of bytes we've probably managed to get away
with this memory corruption.

Change-Id: I23e064edfb8bc9ae156f0fdb78940c6ef1cf47b9
Reviewed-on: http://review.couchbase.org/101548
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>