MB-31797: GetItemsToPersist_Limit does not inc BackfillQSize
Correct the bugs currently in the code that are found by setting
the CB_DEVELOPMENT_ASSERTS flag to ON.

The GetItemsToPersist_Limit test manually pushes items to the
backfill queue without incrementing the backfillQueueSize.
BackfillQueueSize is a NonNegativeCounter so this needs to
be corrected as when we later call getItemsToPersist we decrement
the count and would do so below 0.

Change-Id: I04d10c7998d5a4012e1d30e7a3507540881f0ace
Reviewed-on: http://review.couchbase.org/101549
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>