MB-31797: Make RocksDB assume deletions are for non-existing items
Correct the bugs currently in the code that are found by setting
the CB_DEVELOPMENT_ASSERTS flag to ON.

Modify the RocksDB commit callback to assume that any deletions are
for items that do not exist. This ensures that we overcount the
vBucket document counts under RocksDB (we won't decrement for any
deletions) as opposed to undercount it. This is required as this calls
the PersistenceCallback which would attempt to decrement the vBucket doc
counter to a value less than 0. Skip the now failing RocksDB tests that
relied on this vBucket document counter, or the other counters in the
vBucket that are now not set correctly due to the return value of the
RocksDB commitCallback.

Change-Id: Idc40e14d9f343f91042a2f926f88f98004454ecb
Reviewed-on: http://review.couchbase.org/101551
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>