MB-31910: Initialise phase if hash table eviction policy changes
With the introduction of the new hash table eviction policy (hifi_mfu)
we provided the ability to switch back to the old policy (2-bit_lru) in
case an issue was found in the field.

Unfortunately although the switch can be made if the node is restarted
it currently cannot be done at runtime because the ItemPager phase
will be wrong.

The valid phases for the hifi_mfu eviction policy are REPLICA_ONLY and
ACTIVE_AND_PENDING_ONLY.  The valid phases for the old 2-bit_lru policy
are PAGING_UNREFERENCED and PAGING_RANDOM

This patch re-initialises the ItemPager phase if the hash table eviction
policy has changed.

Change-Id: Ifd1e610db1685dfe6eccb29dd2b78f0dbd516bd7
Reviewed-on: http://review.couchbase.org/101553
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>