MB-31797: Correct VBucketFullEvictionTest MB_31037 stat tracking
Correct the bugs currently in the code that are found by setting
the CB_DEVELOPMENT_ASSERTS flag to ON.

The MB_30137 test does not increment the dirtyQueueSize or
diskQueueSize correctly. As these stats are NonNegativeCounters we
need to set them correctly. To correct this issue we can simply create
the persistence cursor and step it along with the test. This ensures
that when we queue an item in the checkpoint we will set the stats
correctly as Checkpoint::queueDirty will return PERSIST_AGAIN. This
is how we should behave in production environments.

Change-Id: I89da79ef5672fda86c0d10d17cb5c6d4a03f6713
Reviewed-on: http://review.couchbase.org/101550
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>