Add support for Flex Extras in set/add/del with meta
The "unknown command framework" will generate the appropriate
error packet to send back to the client if the handler returns
something else than ENGINE_SUCCESS. Utilize this instead
of generating the error packets inside ep-engine.

Fixed a problem in setWithMeta where we could leak memory
if the first call to setWithMeta returns EWOULDBLOCK and
the second one would fail (not all of the error paths
released the memory).

Use the Request class to fetch the various segments
of the packet to ensure that we correctly jump past the
flex info

Change-Id: I91e70b8aaff392aaa544dfb9a86d741f866e4349
Reviewed-on: http://review.couchbase.org/101689
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>