MB-31811: Add collections support to GET_ALL_VB_SEQNOS
Add the ability to request the high seqno for a given collection
using the GET_ALL_VB_SEQNOS command.

MB-32728: GET_ALL_VB_SEQNOS returns defaultC to old clients

If a client has not enabled collections, then they should be returned
the high seqno of the default collection for the given vBucket
regardless of whever or not they have specified a collection.

Change-Id: I797653fb9ae23005df591246ded32f4c63cf86c5
Reviewed-on: http://review.couchbase.org/102380
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>