MB-31811: Accept 0 as a valid vbucket_state_t in GET_ALL_SEQNOS
To add a collection ID to the GET_ALL_VB_SEQNOS command we need to
add an additional optional field. As vbucket_state_t and
CollectionIDType are both the same length we need to include both values
when we wish to filter GET_ALL_VB_SEQNOS by collection to avoid API
breaking changes. To allow users to filter GET_ALL_VB_SEQNOS for any
vbucket_state_t and a given collection we can make a simple change to
the command to allow a 0 value to return all seqnos for"alive" vBuckets.
This is in keeping with the current default(no vbucket_state_t supplied)
which returns seqnos for vBuckets in the alive, replica, and pending
states.

Make the changes required to accept a 0 value in GET_ALL_VB_SEQNOS
and create appropriately named const values for readability.

Change-Id: Iae2d64055581961721cb247e13637c42eff662fd
Reviewed-on: http://review.couchbase.org/102379
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>