MB-31971: Copy flatbuffers schema into a generated .cc file
The flatbuffers JSON generation code requires that the
schema be provided, which in the example code is done by
loading the fbs file at runtime

We would like to use the JSON generation API but without
having a runtime dependency on the fbs file. The solution
here is to embed the schema file data as a string into
a cc file, users can build the .cc and extern the string

To achieve this (and ensure updates to the fbs result in
a new generated file) we use add_custom_command
to shell out to cmake in script mode (-P) which allows
us to use file command and configure_file at the make
stage.

Change-Id: Ifd95424dee401ece98849d0c8217037bc9ad82b3
Reviewed-on: http://review.couchbase.org/101635
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>