Refactor: Remove isReplication flag and instead rely on vbucket state
We pass the isReplication flag through EventuallyPersistentEngine ->
KVBucket -> VBucket to determine the threshold for deciding whether
there is sufficient memory to allocate a StoredValue, (see
VBucket::hasMemoryForStoredValue)  The flag selects a different
threshold depending on whether we are allocating items received via a
passive stream or not.

Given that the destination of a passive stream cannot be an active
bucket (see DcpConsumer::addStream), in most cases basing the decision
on whether the vbucket is active or not, is equivalent.

There are two exceptions:

1) EPDiskRollbackCB::callback and EPBucket::rollbackUnpersistedItems use
setFromInternal, which uses the active threshold.

2) During warmup when items are added to replica buckets it uses the
active threshold.

To support these two exceptions an UseActiveVBMemThreshold flag is added
to VBucket::hasMemoryForStoredValue.

Change-Id: I3c64ef177ac60a09d602cf650862858f1a965a69
Reviewed-on: http://review.couchbase.org/101640
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>