Revert "MB-31664: Switch over to not using backfill queue"
This reverts commit 5968b4beaae496e080cbf174259ef01a7dabdabc.

By enqueue the backfill snapshot into the CheckpointManager we may end
up OOM and never recover from there. That is because at Replica we close
the open checkpoint (ie, we make it eligible for removing/deallocating)
only when we receive the snapshot-end mutation.
So, if disk-snapshot-size > bucket-quota (which is a usual scenario)
then we can't recover from OOM at Replica.

Change-Id: I56d30056685fc6a7dd4d6c600189a3c5dbbcd36d
Reviewed-on: http://review.couchbase.org/101684
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>