MB-32001: FrameInfo encoding of Durability Requirement
The FrameInfo encding for Durability Requirements looks like:

    Byte/     0           |
       /                  |
      |0 1 2 3 4 5 6 7    |
      +-------------------+
     0|  ID:1 | Len:1 or 3|

The size of the durability requirement is variable length. The first byte
contains the durability level by using the following table:

    0x01 = majority
    0x02 = majority and persist on master
    0x03 = persist to majority

The (optional) 2nd and 3rd byte contains the timeout specified in
milliseconds (network byte order). If omitted the default
timeout value configured on the server will be used.

Change-Id: I1eae2f524c03f4ce5630c3e6310c439c876d6358
Reviewed-on: http://review.couchbase.org/101690
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>