Refactor: Remove checkpoint offset and ckptMetaItemsRead
Instead of keeping cached values of the number of items that have
already been read, (which are used to calculate the number of items that
are the cursor has left to consume - see getNumItemsForCursor), we
calculate dynamically.

This allows us to remove the offset and ckptMetaItemsRead parameters
from CheckpointCursor, which significantly simplifies the code.

Change-Id: I0d9b5c1c6b29e9df3046c3ed8cc7d8af7b71640e
Reviewed-on: http://review.couchbase.org/101810
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>