Refactor: Removes need to record cursor names in a checkpoint
Replaces the list of cursor names used to record the current cursors
in the checkpoint, with a non negative count of the number of
cursors in the checkpoint.

Removes removeInvalidCursorsOnCheckpoint function from CheckpointManager,
which used the list of cursor names.  We never have invalid
cursors in a checkpoint as they are accurately added and removed.

Re-writes the getListOfCursorsToDrop function from CheckpointManager
to not require the list of cursor names.  The re-write of also removes
the need to use the isEligibleToBeUnreferenced function from Checkpoint
which checked if a checkpoint contained the persistent cursor by
iterating through the list of cursor names comparing with the name
"persistence".

Change-Id: Ie81279a43deeffc181248e9c46eef181321e642b
Reviewed-on: http://review.couchbase.org/101852
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>