MB-32078 Return success when resetting append statemachine
When the engine returns KEY_EEXISTS during the storeItem phase of the
append statemachine, and no cas was provided for the operation, the
statemachine should be reset and the operation retried. While the
storeItem function does reset the statemachine, it currently then
returns KEY_EEXISTS. This breaks the processing loop, returning
KEY_EEXISTS to the client instead of retrying the operation.

Change-Id: I9c8fdf637507c98ab7bc1f94f2ee265f7e331d5e
Reviewed-on: http://review.couchbase.org/101959
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>