MB-32123: Fix intermittent failure of "MB-26907 backfill expired value"
Test was intermittently failing due to waiting for a stat to increase
to 5, that had already increased by 5 but due to having two calls of
the test function, the stat does not get reset and gets stuck at 10.
This patch separates the two function calls into their own test calls
in an attempt to resolve this.

Change-Id: I1a42a801619da9f4065af5a4c501f48d1f2c0853
Reviewed-on: http://review.couchbase.org/102058
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>