MB-32591: Change Iterator from a list to a dequeue
To help reduce memory usage we want to be able to remove items from the
checkpoint queue where all cursors are to the right (i.e. the items have
been processed).  This can be achieved more easily when using a deque.

As part of the patch de-duplicate items are no longer removed from the
checkpoint queue, but instead they are set to a nullptr and the entry
in the checkpoint remains so we do not require any special iterator
manipulation.  When iterating (for example when flushing to disk) we
simply skip past any entries in the checkpoint queue that have a
nullptr.

Change-Id: Ia63aef3cde88354e3a990ebdd5e80980e9d81a3b
Reviewed-on: http://review.couchbase.org/102060
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Daniel Owen <owend@couchbase.com>