MB-32117: Introduce the DurabilityMonitor
The DurabilityMonitor (DM) drives the finalization (commit/abort) of a
SyncWrite request.

For doing that, the DM tracks pending SyncWrites in the HashTable and
CheckpointManager and monitors inbound replica ACKs to verify if the
Durability Requirement of tracked SyncWrites is satisfied.

Note that this very first implementation represents what we name
ActiveDurabilityMonitor in the specification.

Change-Id: I03809c9ca62becbba3400a85c1899e30062195c1
Reviewed-on: http://review.couchbase.org/102071
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>