Refactor: Move dcp_last_ variables to producers class
Whenever tests want to check the DCP producer state there's
a bit of boiler plate in extern linking to the mock_dcp.cc
'global' variables. Also, they're globals which limit the
tests we could write.

We can move the variables to be owned by the MockDcpMessageProducer
to fix the problems above. Tests can now be written to use multiple
producers and further work could be done to associate the variables
with streams for a more complete DCP test structure.

Change-Id: Ib9ef225cfff23b6928649756adcda9ceb5268a7b
Reviewed-on: http://review.couchbase.org/102064
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>