MB-32149: Remove single bgfetcher and associated code
The single bgfetcher task is not used for any of our supported storage
backends - they all support "efficient get" and hence we use the
multiBGfetcher task to fetch all outstanding keys for a given vBucket
at once.

As such the single bgfetcher and associated code is dead. Remove it to
simplify handling of background fetching. This included the
bgFetchDelay functionality which was only implmeneted in the
SingleBGFetcher.

Change-Id: I6be9d3eaa79a9d99513105e6cc8731d8016b87e9
Reviewed-on: http://review.couchbase.org/102146
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>