MB-30624: Collection key-length validation
Background/Requirements:

A bucket has historically permitted keys from 1 to 250 bytes, a K/V
operation found with keylen field outside of that range is illegal and
would be failed.

With collections we have to assume that 250 byte keys may be present in
existing buckets and we must also support 'legacy' applications that
read/write the _default collection.

We also need to ensure that collection aware clients can read/write the
default collection by encoding its ID into the key-bytes in the same
way a client encodes and collection-ID (as a unsigned leb128 prefix in
the key-bytes).

Changes:

Firstly to support collection aware clients accessing the default
collection we have to permit a key-length range between 1 to 251, this
ensures the collection aware client can access every key that a legacy
client can.

Secondly we know that the collection-ID when encoded as unsigned leb128
can take between 1 to 5 bytes depending on the value, so we could just
raise the max length of a collection aware client to 255, which is the
absolute maximum we could accept with the 1-byte memcached protocol
keylen field.

However for future proofing, collections does not wish to exhaust all
available bits of the keylen field, allowing for future code to have
some kind of escape in the field if ever needed.

The following rules will apply:

* keylen: the value found in the 1 byte memcached protocol field.
* key: all of the bytes keylen describes.
* collection-ID: the variable length prefix encoding the collection.
* logical-key: the document key, value after removing the collection-ID

If a client does not enable collections with HELLO:

* keylen can be 1 to 250 bytes and all K/V operations are forced into
the  _default collection.

If a client enables collections with HELLO:

* keylen must always be a minimum of 2 and a maximum of 251.
* if the collection-ID is the _default collection, the logical-key max
length of 250 is allowed, ensuring all legacy keys are still reachable.
* if the collection-ID is not the _default collection the logical-key
max length is 246 bytes

The result is that keylen values 252, 253, 254 and 255 are unused and
reserved for future use.

Note1: the reason the logical key length is restricted to a maximum of
246 is to ensure that an application can always encode its longest key
regardless of how many bytes the collection-ID consumes.

Note2: We assume clients will encode the smallest possible _default
collection value, a single 0 byte and not 0x80 00 or longer. Any client
encoding a longer collection-ID will just be limiting their logical-key
space.

Note3: A logical-key length of 0 is invalid and also checked for, this
happens when the entire key is consumed by a leb128 encoding.

Change-Id: Iac260244684f600beecc310ab30a2321ac78dbc4
Reviewed-on: http://review.couchbase.org/102145
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>