Refactor: Update StoredValue bits & bits2 comment clarity
Inside StoredValue, bits had specific member comments inside the
main bits comment which made it unclear whether it applied to the
whole set of members. These have been moved to above their
linking index to improve this clarity.
bits2 has also been updated to clear any confusion that it was
entirely or logically a part of OrderedStoredValue, which is
incorrect.

Change-Id: If9be682e8738e2b5f3c9e75480ef4f8d0e65a562
Reviewed-on: http://review.couchbase.org/102157
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>