MB-32108: Refactor: Return HBL from HashTable::find()
HashTable::find() is used to locate an item in the HashTable when it
hasn't already been locked - it looks up and acquires the appropriate
lock internally, finds the item and returns a pointer to it.

However find()'s current use is limited by the fact it releases the
hash bucket lock as soon as it returns - which means that the returned
pointer isn't necessarily safe to dereference as another thread could
have modified or even removed it. As such it's use is mostly limited
to test code.

To increase the usefulness of this method, return the HashBucketLock
alongside the found StoredValue; this means the returned StoredValue
/can/ be safely be used as long as the HashBucketLock is in scope.

Change-Id: I472e1b939633ccc8c72b0bdb7a5c2c8ec410e79c
Reviewed-on: http://review.couchbase.org/102335
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>