MB-31071: Fix audit of illegal packet
If completely rubbis was sent to memcached we would throw
an exception while trying to fetch the packet to dump, as the
length fields in the packet did not add up to a complete
packet.

Tweak the code so that the content to be dumped is provided
to the audit function, as we know how much data we've got
available when the audit method is called.

Only dump the first 256 bytes of the packet, but add to
the audit trail the number of bytes dropped.

Change-Id: I78863dfb40a42862e8aa176ff6df91a2358289a0
Reviewed-on: http://review.couchbase.org/102367
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>