MB-32083: Fix intermittent failure of vbucket compact test
Fix an issue where we rely on the expiry pager to expire the key,
which can sometimes not run within the duration of the test. This
is fixed by adding another touch of the document which, as it is
now logically expired, will fail and cause the item to be expired
via access.

Change-Id: I04ddad0c9ad9636f2a3eedd9de2f8eecfb173eb2
Reviewed-on: http://review.couchbase.org/102370
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>