MB-32108: HashTable: introduce Committed & Pending perspectives
Background
==========

To support Synchronous writes; the HashTable is modified to support
*Pending* items in addition to the previously supported "normal"
items; which are now termed Committed items.

Committed item represent the current, extenally-visible value of a key.

Pending items are needed to support SyncWrites - they are items which are
proposed to be the "next" value for a key, but are awaiting confirmation
from sufficient other nodes before they can be committed - at which point
the Pending Item is converted to a Committed one.

Different uses of the HashTable will want to access potentially
different items:

  - when performing a read() then only Committed items will typically
    want to be searched for - even if a Pending item exists it should
    be ignored.

  - When performing a write() then both Committed and Pending items
    typically want to be searched - as a Pending item will typically
    block any changes to the key until it is later committed (or
    aborted).

To facilitate this, find() operations on the HashTable expose two
_perspectives_ which items can be searched via:

A) findForRead() - uses Committed perspective. Only returns Committed
   items; Pending items are ignored.

B) findForWrite() - used Pending perspective. Returns both Committed
   and Pending items; if both Pending and Committed items exist for a
   key then the Pending item is returned.

Details
=======

This patch adds initial support for accessing via the two
perspectives; including ensuring that keys with Pending items cannot
be changed via normal mutation methods (only commit() / abort() which
are still to come). Includes unit tests for the new functionality.

Change-Id: I215e476b93c2a3fd7a9b8e58b9a84cc65ad430d5
Reviewed-on: http://review.couchbase.org/102389
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>